requirements.txt
uvicorn
starlette
asyncpg
itsdangerous
psycopg2
ruamel.yaml
requests_oauthlib